Creating a custom-hook will pass the useEffects
